+++ src/util/changelog.ts
@@ -1,13 +1,26 @@
 import * as vscode from "vscode"
 import * as semver from "semver"
 
+/**
+ * displays the changelog using a webview only if the extension was updated and showChangelogOnUpdate is true.
+ *
+ * @param showChangelogOnUpdate Whether or not to show the changelog when the extension is updated
+ * @param context VSCode extension context
+ */
 export function displayChangelog(showChangelogOnUpdate: boolean, context: vscode.ExtensionContext) {
-    if (extensionWasUpdated(showChangelogOnUpdate, context)) {
+    if (showChangelogOnUpdate && extensionWasUpdated(context)) {
         showChangelog()
     }
 }
 
-function extensionWasUpdated(showChangelogOnUpdate: boolean, context: vscode.ExtensionContext): boolean {
+/**
+ * extensionWasUpdated uses VSCode's global states to determine if the last version that was run on this
+ * machine was older than the current running extension.
+ *
+ * @param context VSCode extension context, used to remember the last version run on the current machine
+ * @returns Whether or not the extension was updated
+ */
+function extensionWasUpdated(context: vscode.ExtensionContext): boolean {
     const thisExtension = vscode.extensions.getExtension('brenek.kotlin-for-frc')
     if (thisExtension === undefined) {
         console.error("thisExtension was undefined, the changelog will not be displayed.")
@@ -19,13 +32,12 @@ function extensionWasUpdated(showChangelogOnUpdate: boolean, context: vscode.Ext
 
     context.globalState.update("lastInitVersion", currentVersion)
 
-    if (!showChangelogOnUpdate) {
-        return false
-    }
-
     return semver.satisfies(currentVersion, `>${storedVersion}`)
 }
 
+/**
+ * showChangelog displays the latest release notes (stored in the webviewContent constant) using a VSCode webview panel.
+ */
 export function showChangelog() {
     const panel = vscode.window.createWebviewPanel('kotlin-for-frcChangelog', 'Kotlin For FRC Changelog', vscode.ViewColumn.One, {})
 
@@ -41,23 +53,14 @@ const webviewContent = `<!DOCTYPE html>
 </head>
 <body>
     <h1>Kotlin for FRC Changelog</h1>
-    <h2><a href="https://github.com/BrenekH/kotlin-for-frc/releases/2022.10.1">2022.10.1</a></h2>
+    <h2><a href="https://github.com/BrenekH/kotlin-for-frc/releases/2023.1.1">2023.1.1</a></h2>
     <p><strong>Implemented enhancements:</strong></p>
     <ul>
         <li>
-            <p>Automatically publish builds to <a href="https://open-vsx.org/extension/Brenek/kotlin-for-frc">open-vsx.org</a></p>
+            <p>Update templates for 2023 season</p>
         </li>
         <li>
-            <p>Put <code>wpilibsuite.vscode-wpilib</code> and <code>Brenek.kotlin-for-frc</code> in <code>.vscode/extensions.json</code> as recommended extensions</p>
-        </li>
-        <li>
-            <p>Updated logo to match the new Kotlin logo</p>
-        </li>
-        <li>
-            <p>Switch to <code>@vscode/test-electron</code> testing package (thanks <a href="https://github.com/anuragh2002">@anuragh2002</a>)</p>
-        </li>
-        <li>
-            <p>New shorter extension description</p>
+            <p>Change target GradleRIO version and year to 2023</p>
         </li>
     </ul>
 </body>