+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Kotlin for FRC Changelog
 
+## [2023.1.1](https://github.com/BrenekH/kotlin-for-frc/releases/2023.1.1)
+
+**Implemented enhancements:**
+
+- Update templates for 2023 season
+
+- Change target GradleRIO version and year to 2023
+
 ## [2022.10.1](https://github.com/BrenekH/kotlin-for-frc/releases/2022.10.1)
 
 **Implemented enhancements:**